+++ src/app/shared/cards/book-card/book-card.component.html
