+++ src/app/shared/cards/book-card/book-details/book-details.component.html
