+++ src/app/shared/cards/book-card/book-image/book-image.component.scss
