+++ src/app/shared/cards/book-card/like/like.component.scss
