+++ src/app/shared/cards/genre-card/genre-card.component.html
