+++ src/app/shared/cards/genre-card/genre-card.component.scss
@@ -1,4 +1,4 @@
-@use "../../../assets/styles/variables" as var;
+@use "../../../../assets/styles/variables" as var;
 
 .genre-card {
   width: 210px;
