+++ src/app/shared/cards/genre-card/genre-card.component.ts
